feat: add event name counter; add defaults to event settings; adjust styles and text in event settings

+++ src/services/event/event.service.ts
@@ -855,6 +855,14 @@ export const deleteEvent = async (
     return deleteSessionSettingsResult;
   }
 
+  const deleteAdvancedSettingsResult = await deleteEventAdvancedSettings(
+    eventId,
+  );
+
+  if (!deleteAdvancedSettingsResult.success) {
+    return deleteAdvancedSettingsResult;
+  }
+
   const { error } = await supabase.from("event").delete().eq("id", eventId);
 
   if (error) {
